Self-Adjusting Table
In this project, the objective is control the high of an adjustable table, using computational vision and serial communication. This script was created to run inarduino, to control the position of a stepper motor. Variating the position of the motor shaft, the table would the table would go up or down.