Create college_parking_control
This code was implemented in arduino and uses some libraries, as well as electronic components. It was idealized in this project the implementation of a vacancy control for the UFSC Blumenau parking lot. The automation process was carried out through the arduino platform, with the aid of a simulator (tinkercad), where it was possible to carry out tests. Parking spaces are controlled individually using ultrasonic sensors. The function of the Arduino is to receive the response from the sensors, and to count the free spaces, displaying them on a panel, in addition to controlling a gate that prevents the passage of cars when there are no spaces available.