Translation of a Drone
In this script i am programing at Matlab, but could also use GNU Octave.
This project is part of a college assignment, subject ''Control Systems" or "Feedback Systems". 
The objective of this stage was to design a controller that could control the translation of a drone and that respected some initial conditions specified in the task.